backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyjwt==2.9.0
python-dotenv==1.0.1
httpx==0.27.2
selectolax==0.3.21
pydantic==2.9.2
python-multipart==0.0.12
